server-side caching implemented

+++ src/services/postgres/NotesService.js
@@ -7,9 +7,10 @@ const AuthorizationError = require('../../exceptions/AuthorizationError');
 
 class NotesService {
     // Dalam melakukan tugasnya, fungsi verifyNoteAccess mengandalkan fungsi verifyCollaborator yang dimiliki oleh CollaborationsService. Dengan begitu kita perlu menambahkan dependency terhadap CollaborationsService di dalam NotesService.
-    constructor(collaborationService) {
+    constructor(collaborationService, cacheService) {
         this._pool = new Pool();
         this._collaborationService = collaborationService;
+        this._cacheService = cacheService;
     }
 
     // Pertama pada fungsi addNote, tambahkan properti owner pada parameter objek note. Kemudian, sesuaikan kueri dengan menambahkan nilai owner seperti ini:
@@ -35,23 +36,38 @@ class NotesService {
             throw new InvariantError('Catatan gagal ditambahkan');
         }
 
+        // fungsi this._cacheService.delete berfungsi untuk menghapus cache ketika terjadi perubahan data
+        await this._cacheService.delete(`notes:${owner}`);
+
         return result.rows[0].id;
     }
 
     // Fungsi getNotes tidak akan mengembalikan seluruh catatan yang disimpan pada tabel notes. Melainkan hanya catatan yang dimiliki oleh owner saja.
     async getNotes(owner) {
-        const query = {
-            text: `SELECT notes.* FROM notes
-            LEFT JOIN collaborations ON collaborations.note_id = notes.id
-            WHERE notes.owner = $1 OR collaborations.user_id = $1
-            GROUP BY notes.id`,
-            // Pada kueri di atas, kita menggunakan LEFT JOIN karena tabel notes berada di posisi paling kiri (dipanggil pertama kali). Kueri di atas akan mengembalikan seluruh nilai notes yang dimiliki oleh dan dikolaborasikan dengan owner. Data notes yang dihasilkan berpotensi duplikasi, sehingga di akhir kueri, kita GROUP nilainya agar menghilangkan duplikasi yang dilihat berdasarkan notes.id.
-            values: [owner],
-        };
-        // Di dalamnya kita dapatkan seluruh data notes yang ada di database dengan query “SELECT * FROM notes”.
-        const result = await this._pool.query(query);
-        // Kembalikan fungsi getNotes dengan nilai result.rows yang telah di mapping dengan fungsi mapDBToModel.
-        return result.rows.map(mapDBToModel);
+        try {
+            // mendapatkan catatan dari cache
+            // dalam menetapkan nilai key di cache, kita menggunakan userId (owner) dengan prefix notes:. Dengan begitu, cache akan menyimpan data catatan setiap pengguna secara unik dan terhindar dari tumpang tindih.
+            const result = await this._cacheService.get(`notes:${owner}`);
+            return JSON.parse(result);
+        } catch (error) {
+            // bila gagal, diteruskan dengan mengambil catatan dari database
+            const query = {
+                text: `SELECT notes.* FROM notes
+                LEFT JOIN collaborations ON collaborations.note_id = notes.id
+                WHERE notes.owner = $1 OR collaborations.user_id = $1
+                GROUP BY notes.id`,
+                // Pada kueri di atas, kita menggunakan LEFT JOIN karena tabel notes berada di posisi paling kiri (dipanggil pertama kali). Kueri di atas akan mengembalikan seluruh nilai notes yang dimiliki oleh dan dikolaborasikan dengan owner. Data notes yang dihasilkan berpotensi duplikasi, sehingga di akhir kueri, kita GROUP nilainya agar menghilangkan duplikasi yang dilihat berdasarkan notes.id.
+                values: [owner],
+            };
+            const result = await this._pool.query(query);
+            
+            const mappedResult = result.rows.map(mapDBToModel);
+
+            // Catatan akan disimpan pada cache sebelum fungsi getNotes dikembalikan
+            await this._cacheService.set(`notes:${owner}`, JSON.stringify(mappedResult));
+
+            return mappedResult;
+        }
     }
 
     // Untuk proses pengecekan apakah catatan dengan ID yang diminta adalah hak pengguna, menggunakan fungsi baru yaitu verifyNoteOwner. Fungsi tersebut nantinya akan digunakan pada NotesHandler sebelum mendapatkan, mengubah, dan menghapus catatan berdasarkan id.
@@ -100,7 +116,7 @@ class NotesService {
         // lakukan query untuk mengubah note di dalam database berdasarkan id yang diberikan.
         const updatedAt = new Date().toISOString();
         const query = {
-            text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id',
+            text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner',
             values: [title, body, tags, updatedAt, id],
         };
 
@@ -109,13 +125,17 @@ class NotesService {
         // periksa nilai result.rows bila nilainya 0 (false) maka bangkitkan NotFoundError.
         if (!result.rows.length) {
             throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan');
-        } 
+        }
+        
+        // fungsi this._cacheService.delete berfungsi untuk menghapus cache ketika terjadi perubahan data 
+        const { owner } = result.rows[0];
+        await this._cacheService.delete(`notes:${owner}`);
     }
 
     async deleteNoteById(id) {
         // Lakukan query untuk menghapus note di dalam database berdasarkan id yang diberikan.
         const query = {
-            text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+            text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
             values: [id],
         };
 
@@ -125,6 +145,9 @@ class NotesService {
         if (!result.rows.length) {
             throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan');
         }
+
+        const { owner } = result.rows[0];
+        await this._cacheService.delete(`notes:${owner}`);
     }
 
     // fungsi ini digunakan untuk bertujuan memverifikasi hak akses pengguna (userId) terhadap catatan (id), baik sebagai owner maupun collaboration. Dalam proses verifikasi, fungsi ini tidak melakukan kueri secara langsung ke database. Melainkan ia memanfaatkan fungsi yang sudah dibuat sebelumnya, yakni verifyNoteOwner dan verifyCollaborator.
